% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom.subset.R
\name{geom.subset}
\alias{geom.subset}
\title{Boundary Box}
\usage{
geom.subset(spgeom, top, left, bottom, right)
}
\arguments{
\item{spgeom}{within the for-loop working through movements, use the details from the current movement, which will be described as a row in a dataframe.}

\item{top}{does the movement match any cows? If these cows are to be shifted, stats about the cows are can be calcuated if TRUE. Default is TRUE.}

\item{left}{what are the weights of the cows that will be shifted in this movement? Format is vector of numbers.}

\item{bottom}{Leave a comment describing the success or failure of the movement. Format is a string.}

\item{right}{Leave a comment describing the success or failure of the movement. Format is a string.}
}
\description{
This function intersects a boundary box with a spatial object
}
\examples{
inter = geom.subset(bikeways2015, -27.460070, 153.020613, -27.480243, 153.030741)
     leaflet() \%>\% addTiles() \%>\% addPolylines(data=inter)
}
\keyword{boundary,}
\keyword{box}

