% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SimpleSpatialLinesDataFrame.R, R/SimpleSpatialPointsDataFrame.R, R/SimpleSpatialPolygonsDataFrame.R, R/generics.R
\name{get.zoom.level.SimpleSpatialLinesDataFrame}
\alias{get.zoom.level}
\alias{get.zoom.level.SimpleSpatialLinesDataFrame}
\alias{get.zoom.level.SimpleSpatialPointsDataFrame}
\alias{get.zoom.level.SimpleSpatialPolygonsDataFrame}
\title{Get zoom level from object}
\usage{
\method{get.zoom.level}{SimpleSpatialLinesDataFrame}(x, zoom.level)

\method{get.zoom.level}{SimpleSpatialPointsDataFrame}(x, zoom.level)

\method{get.zoom.level}{SimpleSpatialPolygonsDataFrame}(x, zoom.level)

get.zoom.level(x, zoom.level)
}
\arguments{
\item{x}{object}

\item{zoom.level}{zoom level}
}
\description{
This gets the zoom level
}

