% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leafier-package.R
\docType{package}
\name{leafier}
\alias{leafier}
\alias{leafier-package}
\title{leafier: Render large spatial datasets in leaflet}
\description{
Uses smart analysis to easily plot the visible data at an efficient scale.
}
\section{processing functions}{

The processing functions ...
}

\section{plotting functions}{

The plotting functions ...
}

