% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess objects}
\usage{
preprocess(x, zoom.level = 1:12, simplify.tolerance = 1:12)
}
\arguments{
\item{x}{\code{SpatialPointsDataFrame},  \code{SpatialLinesDataFrame}, \code{SpatialPolygonsDataFrame} object.}

\item{zoom.level}{\code{numeric} with zoom levels at which to pre-process data. Defaults to \code{1:12}.}

\item{simplify.tolerance}{\code{numeric} tolerance to simplify the data at the specified zoom levels. Defaults to \code{c(1,2,3,4,5,6,7,8,12)}.}
}
\value{
\code{SimpleSpatialPointsDataFrame},  \code{SimpleSpatialLinesDataFrame}, \code{SimpleSpatialPolygonsDataFrame} object.
}
\description{
This function preprocesses \code{Spatial*} objects.
}

