% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/render.R
\name{render}
\alias{render}
\alias{render.SimpleSpatialLinesDataFrame}
\alias{render.SimpleSpatialPointsDataFrame}
\alias{render.SimpleSpatialPolygonsDataFrame}
\title{Render objects}
\usage{
render(map, x, bbox, zoom.level, ...)

\method{render}{SimpleSpatialPointsDataFrame}(map, x, bbox, zoom.level, ...)

\method{render}{SimpleSpatialLinesDataFrame}(map, x, bbox, zoom.level, ...)

\method{render}{SimpleSpatialPolygonsDataFrame}(map, x, bbox, zoom.level, ...)
}
\arguments{
\item{x}{\code{SpatialPointsDataFrame},  \code{SpatialLinesDataFrame}, \code{SpatialPolygonsDataFrame} object.}

\item{bbox}{\code{numeric} vector with bounding box coordinates.}

\item{zoom.level}{\code{numeric} with current zoom level in map.}

\item{...}{arguments passed to \code{\link[leaflet]{addMarkers}}, \code{\link[leaflet]{addPolyLines}}, and \code{\link[leaflet]{addPolygons}}.}
}
\value{
\code{SimpleSpatialPointsDataFrame},  \code{SimpleSpatialLinesDataFrame}, \code{SimpleSpatialPolygonsDataFrame} object.
}
\description{
This function renders \code{Spatial*} objects.
}

