% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleSpatialLinesDataFrame.R
\name{SimpleSpatialLinesDataFrame}
\alias{SimpleSpatialLinesDataFrame}
\title{Create SimpleSpatialLinesDataFrame object}
\usage{
SimpleSpatialLinesDataFrame(data, polygons, zoom.level, simplify.tolerance)
}
\description{
This function creates a new \code{SimpleSpatialLinesDataFrame} object.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{data.frame}.}

\item{\code{polygons}}{\code{list} of \code{SpatialPolygons} objects.}

\item{\code{zoom.level}}{\code{numeric} vector of zoom levels.}

\item{\code{simplify.tolerance}}{\code{numeric} vector simplification tolerance for each zoom level.}
}}

