% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleSpatialPolygonsDataFrame.R
\name{SimpleSpatialPolygonsDataFrame}
\alias{SimpleSpatialPolygonsDataFrame}
\title{Create SimpleSpatialPolygonsDataFrame object}
\usage{
SimpleSpatialPolygonsDataFrame(data, polygons, zoom.level, simplify.tolerance)
}
\description{
This function creates a new SimpleSpatialPolygonsDataFrame object
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{data.frame}.}

\item{\code{polygons}}{\code{list} of \code{SpatialPolygons} objects.}

\item{\code{zoom.level}}{\code{numeric} vector of zoom levels.}

\item{\code{simplify.tolerance}}{\code{numeric} vector simplification tolerance for each zoom level.}
}}

