% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/simplifyData.R
\name{simplifyData}
\alias{simplifyData}
\alias{simplifyData.SpatialLines}
\alias{simplifyData.SpatialPoints}
\alias{simplifyData.SpatialPolygons}
\title{Simplify objects}
\usage{
simplifyData(x, tol)

\method{simplifyData}{SpatialPoints}(x, tol)

\method{simplifyData}{SpatialLines}(x, tol)

\method{simplifyData}{SpatialPolygons}(x, tol)
}
\arguments{
\item{x}{\code{SpatialPoints},  \code{SpatialLines}, \code{SpatialPolygons} object.}
}
\value{
\code{SpatialPoints},  \code{SpatialLines}, \code{SpatialPolygons} object.
}
\description{
This function simplifies \code{Spatial*} objects.
}

